git-svn-id: svn+ssh://software-srn.sandia.gov/svn/mapreduce/trunk@780 3fc257bb-8550-0410-81b3-a5c103da83f7

+++ oink/wordfreq.cpp
@@ -45,7 +45,7 @@ void WordFreq::run()
   int nfiles;
   MapReduce *mr = obj->input(1,read_words,&nfiles);
   uint64_t nwords = mr->kv_stats(0);
-  uint64_t nfiles_all;
+  int nfiles_all;
   MPI_Allreduce(&nfiles,&nfiles_all,1,MPI_INT,MPI_SUM,MPI_COMM_WORLD);
 
   mr->collate(NULL);
